Set park to derived state,del conditions
In order to reduce requests made to API, park data corresponding to the park card the user clicks on, is passed to this component using derived state value from allNationalParks, which is set on init app load. Unneeded conditions (if !park) are deleted, as this will now always be truthy.